How does 18446744073709551615 players sound?

+++ main.go
@@ -47,7 +47,7 @@ func main() {
 	}
 	log.Print("Now listening on ", *flagHostPort)
 
-	config.NumSlots = uint8(*flagNumSlots)
+	config.NumSlots = uint64(*flagNumSlots)
 
 	go func() {
 		for {
